+++ CHANGELOG.md
@@ -14,6 +14,8 @@
 
 - [cursor] Fix crash when providing extra large or 0 sizes for cursor theme
 - [sys] Use pkg-config for compile time linking (fixes FreeBSD build without dlopen)
+- [scanner] Force the use of `proc_macro2` fallback implementation, so that the scanner doesn't panic
+  when ran with `RUSTFLAGS="-Cpanic=abort"`.
 
 ## 0.26.6 -- 2020-05-23